+++ Sources/StreamChatSwiftUI/ChatChannel/Reactions/ReactionsOverlayViewModel.swift
@@ -33,7 +33,10 @@ open class ReactionsOverlayViewModel: ObservableObject, ChatMessageControllerDel
             messageController?.deleteReaction(reaction)
         } else {
             // reaction should be added
-            messageController?.addReaction(reaction)
+            messageController?.addReaction(
+                reaction,
+                enforceUnique: utils.messageListConfig.uniqueReactionsEnabled
+            )
         }
     }