+++ CHANGELOG.md
@@ -3,7 +3,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
-### 🔄 Changed
+### 🐞 Fixed
+- Improved loading of gallery images
 
 # [4.46.0](https://github.com/GetStream/stream-chat-swiftui/releases/tag/4.46.0)
 _December 21, 2023_