+++ CHANGELOG.md
@@ -3,6 +3,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
+### 🔄 Changed
+
+# [4.47.0](https://github.com/GetStream/stream-chat-swiftui/releases/tag/4.47.0)
+_January 09, 2024_
+
 ### ✅ Added
 - Config for highlighted composer border color
 - Config for enforcing unique reactions